qual project
